refactor: remove useless code and rename some methods
