+++ packages/player/src/index.ts
@@ -6,7 +6,6 @@ import {
     fmp,
     isSnapshot,
     classifyRecords,
-    isDev,
     radix64,
     logError
 } from '@timecat/utils'
