+++ packages/timecat/src/index.ts
@@ -1,3 +1,3 @@
-export { Recorder } from '@timecat/record'
+export { Recorder } from '@timecat/recorder'
 export { replay } from '@timecat/player'
 export { exportReplay } from '@timecat/utils'
